requirements.txt
pyTelegramBotAPI==4.15.4
APScheduler==3.10.4
SQLAlchemy==2.0.25
python-dotenv==1.0.1
google-generativeai==0.3.1
requests==2.31.0
beautifulsoup4==4.12.3
Pillow==10.2.0
pytz==2024.1
